uowSequence is now uowId as Long

+++ crabzilla-core/src/main/java/io/github/crabzilla/UnitOfWorkJournal.kt
@@ -2,9 +2,8 @@ package io.github.crabzilla
 
 import io.vertx.core.AsyncResult
 import io.vertx.core.Handler
-import java.math.BigInteger
 
 interface UnitOfWorkJournal {
 
-  fun append(unitOfWork: UnitOfWork, aHandler: Handler<AsyncResult<BigInteger>>)
+  fun append(unitOfWork: UnitOfWork, aHandler: Handler<AsyncResult<Long>>)
 }
