+++ crabzilla-core/src/main/java/io/github/crabzilla/UnitOfWorkRepository.kt
@@ -2,19 +2,18 @@ package io.github.crabzilla
 
 import io.vertx.core.AsyncResult
 import io.vertx.core.Handler
-import java.math.BigInteger
 import java.util.*
 
 interface UnitOfWorkRepository {
 
   fun getUowByCmdId(cmdId: UUID, aHandler: Handler<AsyncResult<UnitOfWork>>)
 
-  fun getUowByUowId(uowId: BigInteger, aHandler: Handler<AsyncResult<UnitOfWork>>)
+  fun getUowByUowId(uowId: Long, aHandler: Handler<AsyncResult<UnitOfWork>>)
 
   fun selectAfterVersion(id: Int, version: Version, aggregateRootName: String,
                          aHandler: Handler<AsyncResult<RangeOfEvents>>)
 
-  fun selectAfterUowId(uowId: BigInteger, maxRows: Int,
+  fun selectAfterUowId(uowId: Long, maxRows: Int,
                        aHandler: Handler<AsyncResult<List<UnitOfWorkEvents>>>)
 
   fun getAllUowByEntityId(id: Int, aHandler: Handler<AsyncResult<List<UnitOfWork>>>)
